requirements.txt
Pillow>=6.2.1
progress==1.5